// src/intl.d.ts
export declare const commonMessages: {
    availability: {
        defaultMessage: string;
        id: string;
    };
    catalog: {
        defaultMessage: string;
        id: string;
    };
    chooseFile: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    channel: {
        defaultMessage: string;
        id: string;
    };
    customApps: {
        defaultMessage: string;
        id: string;
    };
    dashboard: {
        defaultMessage: string;
        id: string;
    };
    demo: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    description: {
        defaultMessage: string;
        id: string;
    };
    descriptionOptional: {
        defaultMessage: string;
        id: string;
    };
    discounts: {
        defaultMessage: string;
        id: string;
    };
    drafts: {
        defaultMessage: string;
        id: string;
    };
    email: {
        defaultMessage: string;
        id: string;
    };
    endDate: {
        defaultMessage: string;
        id: string;
    };
    endHour: {
        defaultMessage: string;
        id: string;
    };
    error: {
        defaultMessage: string;
        id: string;
    };
    firstName: {
        defaultMessage: string;
        id: string;
    };
    generalInformations: {
        defaultMessage: string;
        id: string;
    };
    insufficientPermissions: {
        defaultMessage: string;
        id: string;
    };
    lastName: {
        defaultMessage: string;
        id: string;
    };
    limitReached: {
        defaultMessage: string;
        id: string;
    };
    no: {
        defaultMessage: string;
        id: string;
    };
    optionalField: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    properties: {
        defaultMessage: string;
        id: string;
    };
    readOnly: {
        defaultMessage: string;
        id: string;
    };
    requiredField: {
        defaultMessage: string;
        id: string;
    };
    savedChanges: {
        defaultMessage: string;
        id: string;
    };
    sessionExpired: {
        defaultMessage: string;
        id: string;
    };
    somethingWentWrong: {
        defaultMessage: string;
        id: string;
    };
    startDate: {
        defaultMessage: string;
        id: string;
    };
    startHour: {
        defaultMessage: string;
        id: string;
    };
    status: {
        defaultMessage: string;
        id: string;
    };
    summary: {
        defaultMessage: string;
        id: string;
    };
    translationAttributes: {
        defaultMessage: string;
        id: string;
    };
    unauthorizedDashboardAccess: {
        defaultMessage: string;
        id: string;
    };
    uploadImage: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    yes: {
        defaultMessage: string;
        id: string;
    };
    date: {
        defaultMessage: string;
        id: string;
    };
    time: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    welcome: {
        defaultMessage: string;
        id: string;
    };
};
// ...
export declare const sectionNames: {
    apps: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    attributes: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    categories: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    channels: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    collections: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    configuration: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    customers: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    draftOrders: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    exchangeRates: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    giftCards: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    home: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    navigation: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    orders: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    pageTypes: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    pages: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    permissionGroups: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    plugins: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    productTypes: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    products: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    sales: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    serviceAccounts: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    shipping: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    siteSettings: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    staff: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    taxes: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    translations: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    vouchers: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    warehouses: {
        defaultMessage: string;
        id: string;
        description: string;
    };
    webhooks: {
        defaultMessage: string;
        id: string;
        description: string;
    };
};